IoUring: Move exception handling from CompletionQueue to IoUringIoHandler
Motivation:

The exception handling belongs to the IoUringIoHandler as its the one that "drives" the CompletionQueue.

Modifications:

Move exception handling to IoUringIoHandler

Result:

Cleanup
